Added three new objects to test recognition better
