+++ Assembly-CSharp.csproj
@@ -52,6 +52,10 @@
      <Compile Include="Assets\NewBehaviourScript.cs" />
      <Compile Include="Assets\ObjectTargetPlayAudio.cs" />
      <Compile Include="Assets\Oscillator.cs" />
+     <Compile Include="Assets\Scripts\Armring3DViewer.cs" />
+     <Compile Include="Assets\Scripts\ArmringDetailView.cs" />
+     <Compile Include="Assets\Scripts\MenuScript.cs" />
+     <Compile Include="Assets\Scripts\SceneNavigator.cs" />
      <Compile Include="Assets\ToonBot\ToonBot(Free) Assets\Controller\RobotTestScriptFree.cs" />
      <Compile Include="Assets\Vuforia\Scripts\BackgroundPlaneBehaviour.cs" />
      <Compile Include="Assets\Vuforia\Scripts\CloudRecoBehaviour.cs" />
@@ -107,6 +111,7 @@
      <None Include="Assets\Vuforia\Shaders\ClippingMask.shader" />
      <None Include="Assets\Vuforia\Shaders\BrightTexture.shader" />
      <None Include="Assets\Editor\QCAR\TargetsetData\himba-db_info.xml" />
+     <None Include="Assets\example.txt" />
      <None Include="Assets\Vuforia\Shaders\Text3D.shader" />
  <Reference Include="UnityEngine.UI">
  <HintPath>D:/Programs/Unity/Editor/Data/UnityExtensions/Unity/GUISystem/UnityEngine.UI.dll</HintPath>
