+++ Assets/MockScene.unity.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 06acea87825323d48898209284719f5b
+timeCreated: 1496140944
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
