+++ Assets/Scripts.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 579e687ee1e90954f922aa14793d3b49
+folderAsset: yes
+timeCreated: 1496140695
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
