+++ Assets/example.txt.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 019ae31f55ca1e2428d1351be362b1d8
+timeCreated: 1496140699
+licenseType: Free
+TextScriptImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
