+++ Library/LastSceneManagerSetup.txt
@@ -2,3 +2,6 @@ sceneSetups:
 - path: Assets/basic.unity
   isLoaded: 1
   isActive: 1
+- path: Assets/MockScene.unity
+  isLoaded: 0
+  isActive: 0
