--- Scripts/Armring3DViewer.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: b19ae70eca392bf4f9ea580da725b061
-timeCreated: 1496065454
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
