--- Scripts/ArmringDetailView.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: f171c5be3e9f59f4f94ee65224e9686b
-timeCreated: 1496066687
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
