--- Scripts/MenuScript.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: f1ac6caf63b9b594baead7c723b48d96
-timeCreated: 1495892636
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
