--- Scripts/SceneNavigator.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 3766b0e08adf4704eb70b9ccc920f25f
-timeCreated: 1496063584
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
